Fixed Bullet orientation/rotation, player movement animation direction
I hide the child node 'Trail' of the Player node; this child node do not follow the player rotation and looks wired while Player change direction.
I also changed the project name to "Lightning-game-One-man-army-of-the-space" as everyone have the same project in their computer which might bring confusion. You can change the name if you have a good one.